// backups/20260304_1507/types.ts
export interface Package {
  id: string;
  name: string;
  price: number;
  propertyLimit: number;
  durationDays: number;
  featuredLimit: number;
  superFeaturedLimit?: number;
  description: string;
  isActive?: boolean;
}